Refactor VaToast plugin.
