--- packages/ui/src/components/va-toast/install.ts
@@ -1,8 +1,0 @@
-import { App } from 'vue'
-import Notification from './index'
-
-export default {
-  install (app: App) {
-    app.config.globalProperties.$vaToast = new Notification(app)
-  },
-}
