+++ packages/ui/src/vue-book/book-main.ts
@@ -3,7 +3,7 @@ import App from './BookApp.vue'
 import DropdownPopperSubplugin
   from '../components/va-dropdown/dropdown-popover-subplugin'
 // import ColorHelpersPlugin from '../components/vuestic-utilities/color-helpers-plugin'
-import ToastInstall from '../components/va-toast/install'
+import ToastInstall from '../components/va-toast/plugin'
 
 import { VueBookComponents, createRoute } from 'vue-book'
 import { createRouter, createWebHashHistory } from 'vue-router'
